06-rag-application/Backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain_huggingface
langchain
langchain_core
python-dotenv
sse_starlette
langserve
transformers
huggingface_hub
protobuf
langchain_community
langchain_openai
pinecone[grpc]
GitPython
chromadb
sqladmin